fix: fixes the other fragments opacity
