+++ app/src/main/java/app/mlauncher/style/SettingsTheme.kt
@@ -93,8 +93,8 @@ fun SettingsTheme(
         settings = RoundedCornerShape(CORNER_RADIUS),
     )
     val replacementColor = ReplacementColor(
-        settings = colorResource( if (isDark) R.color.blackTrans80 else R.color.blackInverseTrans80 ),
-        selector = colorResource( if (isDark) R.color.blackTrans80 else R.color.blackInverseTrans80 ),
+        settings = colorResource( if (isDark) R.color.blackTrans50 else R.color.blackInverseTrans50 ),
+        selector = colorResource( if (isDark) R.color.blackTrans50 else R.color.blackInverseTrans50 ),
         border = colorResource( if (isDark) R.color.blackInverseTrans25 else R.color.whiteInverseTrans25 ),
     )
     CompositionLocalProvider(
